server/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-CORS==3.0.10
Flask-SocketIO==5.3.2
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
SQLAlchemy==2.0.4
Werkzeug==2.2.3
eventlet==0.33.3
